fixing problems that occured when the grid amount was changed.
When the gridamount was changed using the size button, cells didn't
change state on click. I made a new function, cellClicks() (all click
events from registerevens that used the divs with .cell are put into
this function instead of registerevents.) Calling this function after
the amount was changed fixed the problem.

Passing in a hash, that defined a grid with another size, caused a lot
of problems (cells didnt change state on click, the game interval acted
weird and the game didn't pause at all.)

By toying around with the code for a long time i fexed the problem.

The code might not be eficient or pretty, but it does work.